+++ gui/widget/RotarySlider.h
@@ -143,7 +143,7 @@ protected:
       g.strokePath(upperRail, strokeType);
     } else {
       const int numSelections = (int)maxValue - (int)minValue;
-      for (int i = 0; i <= numSelections; i++) {
+      for (size_t i = 0; i <= numSelections; i++) {
         const float selectionValue = minValue + i;
         const float rawSelectionAngle = juce::jmap(selectionValue,
                                                    minValue,
